backend/pyproject.toml
[tool.poetry]
name = "mailmind"
version = "0.1.0"
description = "An intelligent email management system"
authors = ["yee"]
readme = "README.md"
packages = [{ include = "app" }, { include = "tests" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
flask = {extras = ["async"], version = "^3.1.0"}
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
werkzeug = "^3.0.1"
gunicorn = "22.0.0"
flask-jwt-extended = "^4.7.1"
flask-cors = "^4.0.0"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
flask-sqlalchemy = "^3.1.1"
google-auth = "^2.28.1"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.120.0"
pytest-mock = "^3.12.0"
beautifulsoup4 = "^4.12.0"
openai = "^1.66.0"
flask-mail = "^0.9.1"
pytz = "^2025.2"
apscheduler = "^3.10.4"
httpx = "^0.26.0"
flask-migrate = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
ruff = "^0.2.1"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "run:main"
dev = "run:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing"
pythonpath = ["."]
